requirements.txt
streamlit>=1.28.0
openai-whisper>=20231117
langchain>=0.1.0
langchain-huggingface>=0.0.6
transformers>=4.35.0
torch>=2.0.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
fpdf>=1.7.2
python-telegram-bot>=20.0
google-auth>=2.23.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.100.0
schedule>=1.2.0
python-dotenv>=1.0.0
cryptography>=41.0.0
googletrans>=4.0.0rc1
accelerate>=0.24.0 